requirements.txt
alpaca-py==0.43.2
annotated-types==0.7.0
certifi==2026.2.25
charset-normalizer==3.4.7
confluent-kafka==2.14.0
idna==3.11
massive==2.5.0
msgpack==1.1.2
numpy==2.4.4
pandas==3.0.2
pydantic==2.13.1
pydantic_core==2.46.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.33.1
six==1.17.0
sseclient-py==1.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0
